Move window state functions to `settings`
It makes more sense for them to be there, as they were the only things
making use of `settings`.

It also cleans `app` a bit.
